requirements/production.txt
-r base.txt
psycopg2
gunicorn
whitenoise
redis
django-redis
# Anything else for production environment
